Added login validation to the rest of the files

+++ php/deleteFile.php
@@ -1,15 +1,16 @@
 <?php
+if ($_SESSION["auth"] == true) {
+    $deleteFile = new DeleteFile();
 
-$deleteFile = new DeleteFile();
+    $deleteFile->delete($_GET["file"]);
 
-$deleteFile->delete($_GET["file"]);
-
-class DeleteFile
-{
-    function delete(String $file)
+    class DeleteFile
     {
-        if (str_contains($file, ".ovpn") || str_contains($file, ".conf")) {
-            unlink("../vpnConfigs/$file");
-        };
+        function delete(String $file)
+        {
+            if (str_contains($file, ".ovpn") || str_contains($file, ".conf")) {
+                unlink("../vpnConfigs/$file");
+            };
+        }
     }
 }
