+++ php/stopAP.php
@@ -1,2 +1,4 @@
 <?php
+if ($_SESSION["auth"] == true) {
     exec("sudo pkill create_ap");
+}
